+++ .eslintrc.js
@@ -14,5 +14,6 @@ module.exports = {
         "comma-dangle": [ "error" , "always-multiline" ],
         "array-bracket-newline": [ "error", "consistent" ],
         "function-paren-newline": [ "error", "multiline" ],
+        "no-control-regex": ["off"],
     },
 };